+++ Models/Calculations/NumberCalculation.cs
@@ -18,7 +18,7 @@ public class NumberCalculation(Calculation? receiver) : Calculation(receiver)
     /// <summary>
     /// 計算の最初の数値を取得、設定します。
     /// </summary>
-    public int Number { get; set; } = 0;
+    public decimal Number { get; set; } = 0;
 
     /// <summary>
     /// 計算した結果を取得、設定します。
@@ -27,5 +27,5 @@ public class NumberCalculation(Calculation? receiver) : Calculation(receiver)
     /// 計算した結果です。次の計算があった場合は、計算対象となります。
     /// nullになりません。
     /// </remarks>
-    public override int? Result => Number;
+    public override decimal? Result => Number;
 }
