+++ Models/NumberToken.cs
@@ -6,10 +6,10 @@
 /// <param name="number">
 /// 文字列で表された数値。
 /// </param>
-public class NumberToken(int number) : Token
+public class NumberToken(decimal number) : Token
 {
     /// <summary>
     /// 入力された数値を取得します。
     /// </summary>
-    public int Number { get; set; } = number;
+    public decimal Number { get; set; } = number;
 }
